source_by_pc supports multiple files

+++ tests/test_contracts/ms/MultiSource.sol
@@ -6,13 +6,13 @@ import {MultiSourceUtils} from "./MultiSourceLib.sol";
 
 contract MultiSource{
     // using MultiSourceLib for uint32;
-    uint32 i;
+    uint256 i;
     MultiSourceUtils utils;
     constructor(){
         utils = new MultiSourceUtils();
     }
 
-    function add(uint32 x, uint32 y) public returns (uint256){
+    function add(uint256 x, uint256 y) public returns (uint256){
         // i = x.unsafeAdd(y);
         i = utils.unsafeAdd(x, y);
         return i;
